flask_backend/pyproject.toml
[tool.poetry]
name = "flask-backend"
version = "0.1.0"
description = "Flask Blog API backend"
authors = ["Marat-77"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Marat-77/flask_blog"
packages = [{include = "flask_backend"}]

[tool.poetry.dependencies]
python = "~3.9"
Flask = "^2.3.2"
flask-sqlalchemy = "^3.0.5"
psycopg2-binary = "^2.9.6"
flask-migrate = "^4.0.4"
flask-restx = "^1.1.0"
flask-jwt-extended = "^4.5.2"
gunicorn = "^20.1.0"
bcrypt-flask = "^1.0.2"
python-dotenv = "^1.0.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
